Fixed Swagger spec (#6253)
* Fixed Swagger spec

In the Swagger spec, the operationId `Conversations_GetConversationMembers` was defined twice, and that caused an issue with the Swagger CLI's validation. One of the two instances should have been `Conversations_GetConversationMember` instead.

* Better compatibility with go-swagger

Adding `"x-nullable": true` to date fields and `"x-omitempty": true` to arrays to ensure that the resulting JSON doesn't include empty dates and empty arrays are not rendered as "null" when using go-swagger to generate a client